// src/lib/types.ts
export interface ModulePermission {
  view: boolean;
  create: boolean;
  edit: boolean;
  delete: boolean;
  export?: boolean;
}

export interface RolePermissions {
  dashboard: {
    view: boolean;
  };
  userManagement: ModulePermission;
  listenerManagement: ModulePermission;
  sessionManagement: ModulePermission & {
    endSession: boolean;
  };
  compliance: ModulePermission & {
    viewMessages: boolean;
    flagContent: boolean;
  };
  walletPayments: ModulePermission & {
    processRefund: boolean;
    approveWithdrawal: boolean;
    manualAdjustment: boolean;
  };
  supportTicketing: ModulePermission & {
    assignTickets: boolean;
    closeTickets: boolean;
  };
  notifications: ModulePermission & {
    sendPush: boolean;
    sendEmail: boolean;
  };
  reports: {
    view: boolean;
    export: boolean;
    accessFinancial: boolean;
  };
  settings: ModulePermission & {
    modifyRazorpay: boolean;
    modifyCommission: boolean;
  };
  adminManagement: ModulePermission;
  rolesPermissions: ModulePermission;
  systemHealth: {
    view: boolean;
  };
}